requirements.txt
moomoo-openapi>=4.0.0
anthropic>=0.20.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
aiohttp>=3.9.0
asyncpg>=0.29.0
python-telegram-bot>=21.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
